requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.2
djangorestframework==3.15.2
timedelta==2020.12.3
djangorestframework-simplejwt==5.4.0
pyjwt==2.10.1